Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 1339-maximum-product-of-splitted-binary-tree/1339-maximum-product-of-splitted-binary-tree.java
@@ -14,27 +14,27 @@
  * }
  */
 class Solution {
-    long maxProd=0;
-    long SUM=0;
-    public long totalSum(TreeNode root){
+    int maxProd=0;
+    int SUM=0;
+    public int totalSum(TreeNode root){
         if(root==null) return 0;
-        long sumLeft=totalSum(root.left);
-        long sumRight=totalSum(root.right);
+        int sumLeft=totalSum(root.left);
+        int sumRight=totalSum(root.right);
 
-        long Sum=root.val+ sumLeft+sumRight;
+        int Sum=root.val+ sumLeft+sumRight;
        
 
         return Sum;
 
     }
-    public long findSum(TreeNode root){
+    public int findSum(TreeNode root){
         if(root==null) return 0;
-        long sumLeft=findSum(root.left);
-        long sumRight=findSum(root.right);
+        int sumLeft=findSum(root.left);
+        int sumRight=findSum(root.right);
 
-        long SubTreeSum=root.val+ sumLeft+sumRight;
+        int SubTreeSum=root.val+ sumLeft+sumRight;
 
-        long remSubTreeSum=SUM-SubTreeSum;
+        int remSubTreeSum=SUM-SubTreeSum;
         maxProd=Math.max(maxProd,SubTreeSum*remSubTreeSum);
 
         return SubTreeSum;
@@ -46,6 +46,6 @@ class Solution {
         SUM=totalSum(root);
         findSum(root);
 
-        return (int)maxProd%1000000007;
+        return (maxProd%1000000007);
    }
 }
